rename packages/tf files
